Moved radio button click handlers to TypePanel
